+++ Source/DynamicScene/DynamicPolyline.js
@@ -1,10 +1,17 @@
 /*global define*/
-define([
-        '../Core/defaultValue',
-        '../Core/defined'],
-function(
+define(['../Core/defaultValue',
+        '../Core/defined',
+        '../Core/defineProperties',
+        '../Core/DeveloperError',
+        '../Core/Event',
+        './createObservableProperty'
+    ], function(
         defaultValue,
-        defined) {
+        defined,
+        defineProperties,
+        DeveloperError,
+        Event,
+        createObservableProperty) {
     "use strict";
 
     /**
@@ -13,64 +20,98 @@ function(
      * @constructor
      */
     var DynamicPolyline = function() {
+        this._color = undefined;
+        this._outlineColor = undefined;
+        this._outlineWidth = undefined;
+        this._show = undefined;
+        this._width = undefined;
+        this._propertyChanged = new Event();
+    };
+
+    defineProperties(DynamicPolyline.prototype, {
+        /**
+         * Gets the event that is raised whenever a new property is assigned.
+         * @memberof DynamicPolyline.prototype
+         * @type {Event}
+         */
+        propertyChanged : {
+            get : function() {
+                return this._propertyChanged;
+            }
+        },
+
         /**
          * Gets or sets the {@link Color} {@link Property} specifying the the line's color.
+         * @memberof DynamicPolyline.prototype
          * @type {Property}
          */
-        this.color = undefined;
+        color : createObservableProperty('color', '_color'),
+
         /**
          * Gets or sets the {@link Color} {@link Property} specifying the the line's outline color.
+         * @memberof DynamicPolyline.prototype
          * @type {Property}
          */
-        this.outlineColor = undefined;
+        outlineColor : createObservableProperty('outlineColor', '_outlineColor'),
+
         /**
          * Gets or sets the numeric {@link Property} specifying the the line's outline width.
+         * @memberof DynamicPolyline.prototype
          * @type {Property}
          */
-        this.outlineWidth = undefined;
+        outlineWidth : createObservableProperty('outlineWidth', '_outlineWidth'),
+
         /**
+         * @memberof DynamicPolyline.prototype
          * Gets or sets the boolean {@link Property} specifying the line's visibility.
          * @type {Property}
          */
-        this.show = undefined;
+        show : createObservableProperty('show', '_show'),
+
         /**
          * Gets or sets the numeric {@link Property} specifying the the line's width.
+         * @memberof DynamicPolyline.prototype
          * @type {Property}
          */
-        this.width = undefined;
-    };
+        width : createObservableProperty('width', '_width')
+    });
 
     /**
-     * Given two DynamicObjects, takes the polyline properties from the second
-     * and assigns them to the first, assuming such a property did not already exist.
+     * Duplicates a DynamicPolyline instance.
+     * @memberof DynamicPolyline
      *
-     * @param {DynamicObject} targetObject The DynamicObject which will have properties merged onto it.
-     * @param {DynamicObject} objectToMerge The DynamicObject containing properties to be merged.
+     * @param {DynamicPolyline} [result] The object onto which to store the result.
+     * @returns {DynamicPolyline} The modified result parameter or a new instance if one was not provided.
      */
-    DynamicPolyline.mergeProperties = function(targetObject, objectToMerge) {
-        var polylineToMerge = objectToMerge.polyline;
-        if (defined(polylineToMerge)) {
-
-            var targetPolyline = targetObject.polyline;
-            if (!defined(targetPolyline)) {
-                targetObject.polyline = targetPolyline = new DynamicPolyline();
-            }
-
-            targetPolyline.color = defaultValue(targetPolyline.color, polylineToMerge.color);
-            targetPolyline.width = defaultValue(targetPolyline.width, polylineToMerge.width);
-            targetPolyline.outlineColor = defaultValue(targetPolyline.outlineColor, polylineToMerge.outlineColor);
-            targetPolyline.outlineWidth = defaultValue(targetPolyline.outlineWidth, polylineToMerge.outlineWidth);
-            targetPolyline.show = defaultValue(targetPolyline.show, polylineToMerge.show);
+    DynamicPolyline.prototype.clone = function(result) {
+        if (!defined(result)) {
+            result = new DynamicPolyline();
         }
+        result.color = this.color;
+        result.width = this.width;
+        result.outlineColor = this.outlineColor;
+        result.outlineWidth = this.outlineWidth;
+        result.show = this.show;
+        return result;
     };
 
     /**
-     * Given a DynamicObject, undefines the polyline associated with it.
+     * Assigns each unassigned property on this object to the value
+     * of the same property on the provided source object.
+     * @memberof DynamicPolyline
      *
-     * @param {DynamicObject} dynamicObject The DynamicObject to remove the polyline from.
+     * @param {DynamicPolyline} source The object to be merged into this object.
+     * @exception {DeveloperError} source is required.
      */
-    DynamicPolyline.undefineProperties = function(dynamicObject) {
-        dynamicObject.polyline = undefined;
+    DynamicPolyline.prototype.merge = function(source) {
+        if (!defined(source)) {
+            throw new DeveloperError('source is required.');
+        }
+        this.color = defaultValue(this.color, source.color);
+        this.width = defaultValue(this.width, source.width);
+        this.outlineColor = defaultValue(this.outlineColor, source.outlineColor);
+        this.outlineWidth = defaultValue(this.outlineWidth, source.outlineWidth);
+        this.show = defaultValue(this.show, source.show);
     };
 
     return DynamicPolyline;