+++ Source/DynamicScene/GeoJsonDataSource.js
@@ -49,7 +49,7 @@ define([
         } else {
             var i = 2;
             var finalId = id;
-            while (defined(dynamicObjectCollection.getObject(finalId))) {
+            while (defined(dynamicObjectCollection.getById(finalId))) {
                 finalId = id + "_" + i;
                 i++;
             }
@@ -405,7 +405,7 @@ define([
             }
         }
 
-        this._dynamicObjectCollection.clear();
+        this._dynamicObjectCollection.removeAll();
 
         var dataSource = this;
         return when(crsFunction, function(crsFunction) {